SnappetChallenge/SnappetChallenge/Converters/IntToStringConverter.cs
using System.Globalization;
using System.Windows.Data;

namespace SnappetChallenge.Converters
{
    public class IntToStringConverter : IValueConverter
    {
        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
        {
            if (value is int intValue)
            {
                return intValue switch
                {
                    0 => "No",
                    1 => "Yes",
                    _ => "Unknown"
                };
            }
            return "";
        }

        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
        {
            throw new NotImplementedException();
        }
    }
}